requirements.txt
aiohttp==3.7.4
argparse==1.4.0
configparser==5.0.2
async-timeout==3.0.1
asyncpg==0.23.0
attrs==21.2.0
chardet==3.0.4
git+git://github.com/Rapptz/discord.py@master#egg=discord.py # Discord.py 2.0 (Master)
humanize==3.10.0
idna==3.2
jishaku==2.0.0
multidict==5.1.0
python-dotenv==0.18.0
pyyaml==5.4.1
six==1.16.0
typing-extensions==3.10.0.0
wheel==0.36.2
yarl==1.6.3